pyproject.toml
[tool.poetry]
name = "python-arango-ogm"
version = "0.1.6"
description = "Python-based package implementing an ORM framework for arango; built on top of python-arango."
authors = ["Steven Miers <steven.miers@gmail.com>"]
readme = "README.md"
license = "The Unlicense"

[tool.poetry.dependencies]
python = "^3.12.2"
python-arango = "^7.9.1"
typer = {extras = ["all"], version = "^0.9.0"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff-lsp = "^0.0.52"
pdoc = "^14.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.1"
pytest-mock = "*"

[tool.poetry.scripts]
pao-migrate = "python_arango_ogm.cli.pao_migrate:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#packages = [
#    { include = "python_arango_ogm" }
#]
